fst-01 restructure project
