+++ backend/app/api/api_v1/endpoints/auth.py
@@ -2,7 +2,7 @@ from fastapi.security import OAuth2PasswordRequestForm
 from fastapi import APIRouter, Depends, HTTPException, status
 from datetime import timedelta
 
-from app.db.session import get_db
+from app.core.session import get_db
 from app.core import security
 from app.core.auth import authenticate_user, sign_up_new_user
