+++ backend/app/api/api_v1/endpoints/tests/test_users.py
@@ -1,4 +1,4 @@
-from app.db import models
+from app.users import models
 
 
 def test_get_users(client, test_superuser, superuser_token_headers):
