+++ backend/app/api/api_v1/endpoints/users.py
@@ -1,15 +1,15 @@
 from fastapi import APIRouter, Request, Depends, Response, encoders
 import typing as t
 
-from app.db.session import get_db
-from app.db.crud import (
+from app.core.session import get_db
+from app.users.crud import (
     get_users,
     get_user,
     create_user,
     delete_user,
     edit_user,
 )
-from app.db.schemas import UserCreate, UserEdit, User, UserOut
+from app.users.schemas import UserCreate, UserEdit, User, UserOut
 from app.core.auth import get_current_active_user, get_current_active_superuser
 
 users_router = r = APIRouter()
