+++ backend/app/arrangment/schemas.py
@@ -1,0 +1,15 @@
+from typing import Optional
+from sqlmodel import Field, SQLModel
+from .models import LocationBase, Location
+
+
+class LocationCreate(LocationBase):
+    pass
+
+
+class LocationRead(LocationBase):
+    id: int
+
+
+class LocationUpdate(SQLModel):
+    name: Optional[str] = None
