+++ backend/app/core/auth.py
@@ -2,9 +2,9 @@ import jwt
 from fastapi import Depends, HTTPException, status
 from jwt import PyJWTError
 
-from app.db import models, schemas, session
-from app.db.crud import get_user_by_email, create_user
-from app.core import security
+from app.users import models, schemas
+from app.users.crud import get_user_by_email, create_user
+from app.core import security, session
 
 
 async def get_current_user(
