+++ backend/app/core/session.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from app.core import config
+from . import config
 
 engine = create_engine(
     config.SQLALCHEMY_DATABASE_URI,
