+++ backend/app/initial_data.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 
-from app.db.session import get_db
-from app.db.crud import create_user
-from app.db.schemas import UserCreate
-from app.db.session import SessionLocal
+from app.users.crud import create_user
+from app.users.schemas import UserCreate
+from app.core.session import SessionLocal
 
 
 def init() -> None:
