--- backend/app/tests/test_tasks.py
@@ -1,6 +1,0 @@
-from app import tasks
-
-
-def test_example_task():
-    task_output = tasks.example_task("Hello World")
-    assert task_output == "test task returns Hello World"
